requirements.txt
aioredis==1.3.1
amqp==5.1.1
anyio==3.5.0
appnope==0.1.3
APScheduler==3.6.3
asgiref==3.6.0
appnope==0.1.3
APScheduler==3.6.3
asgiref==3.4.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
backcall==0.2.0
billiard==3.6.4.0
black==21.12b0
cachetools==5.0.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
dadata==21.10.1
daphne==3.0.2
dateparser==1.1.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.5
distlib==0.3.6
Django==4.0.1
django-channels-jwt-auth-middleware==1.0.0
django-cors-headers==3.11.0
django-crontab==0.7.1
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.1
drf-yasg==1.20.0
exceptiongroup==1.0.4
executing==0.8.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
ipython==8.0.1
itypes==1.2.0
jedi==0.18.1
jet-bridge-base==1.2.0
Jinja2==3.0.3
kombu==5.2.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
msgpack==1.0.3
mypy-extensions==0.4.3
numpy==1.23.5
oauthlib==3.1.1
packaging==21.3
pandas==1.5.2
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
pluggy==1.0.0
prompt-toolkit==2.0.9
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pygame==2.1.2
Pygments==2.11.2
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.6
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.0
python3-openid==3.2.0
pytils==0.4.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.2.2
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sentry-sdk==1.5.12
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
SQLAlchemy==1.4.35
sqlparse==0.4.2
stack-data==0.1.4
tomli==1.2.3
tornado==6.2
traitlets==5.1.1
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
voximplant-apiclient==1.3.2
wcwidth==0.2.5
wrapt==1.14.0
zope.interface==5.4.0
